Added string sorting and made index double

+++ QuickSort/src/quickSort/Sortable.java
@@ -2,9 +2,9 @@ package quickSort;
 
 public class Sortable {
 	private Object o;
-	private int i;
+	private double i;
 	
-	public Sortable(Object object, int index) {
+	public Sortable(Object object, double index) {
 		super();
 		o = object;
 		i = index;
@@ -16,7 +16,7 @@ public class Sortable {
 	public void setObject(Object object) {
 		o = object;
 	}
-	public int getIndex() {
+	public double getIndex() {
 		return i;
 	}
 	public void setIndex(int index) {
